+++ packages/tabs/src/TabTitle.tsx
@@ -111,6 +111,13 @@ const listTitle = css`
   &:hover:after {
     transform: scaleX(0.95);
   }
+
+  &:active:after {
+    &:after {
+      transform: scaleX(1);
+      background-color: ${uiColors.green.base};
+    }
+  }
 `;
 
 interface BaseTabProps {
@@ -122,6 +129,7 @@ interface BaseTabProps {
   children?: React.ReactNode;
   className?: string;
   disabled?: boolean;
+  isAnyTabFocused?: boolean;
 }
 
 const TabTitle: ExtendableBox<BaseTabProps, 'button'> = ({
@@ -132,6 +140,7 @@ const TabTitle: ExtendableBox<BaseTabProps, 'button'> = ({
   ariaControl,
   index,
   darkMode,
+  isAnyTabFocused,
   ...rest
 }: BaseTabProps) => {
   const { usingKeyboard: showFocus } = useUsingKeyboardContext();
@@ -139,10 +148,10 @@ const TabTitle: ExtendableBox<BaseTabProps, 'button'> = ({
   const mode = darkMode ? Mode.Dark : Mode.Light;
 
   useEffect(() => {
-    if (!disabled && selected && titleRef.current) {
+    if (isAnyTabFocused && !disabled && selected && titleRef.current) {
       titleRef.current.focus();
     }
-  }, [disabled, selected]);
+  }, [isAnyTabFocused, disabled, selected, titleRef]);
 
   const sharedTabProps = {
     className: cx(