+++ website/pages/index.tsx
@@ -1,10 +1,5 @@
 import React from 'react';
-import Navigation from 'components/navigation';
 
 export default function Home() {
-  return (
-    <div>
-      <Navigation />
-    </div>
-  );
+  return <div>this is the home page</div>;
 }