+++ packages/portal/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leafygreen-ui/portal
 
+## 3.1.0
+
+### Minor Changes
+
+- 627333c2: Portal can now accept a `null` value for the `container` prop
+
+### Patch Changes
+
+- Updated dependencies [ee7923d3]
+  - @leafygreen-ui/lib@6.1.2
+
 ## 3.0.1
 
 ### Patch Changes