+++ packages/text-area/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @leafygreen-ui/text-area
 
+## 2.0.0
+
+### Major Changes
+
+- 69354cdd: className prop is now spread on container `div` rather than `textarea` element
+
+### Minor Changes
+
+- 627333c2: `State` is now a named export from the package
+
+### Patch Changes
+
+- ee7923d3: Changes how we extend the types of HTMLElements, and standardizes how we document this across readmes
+- Updated dependencies [ee7923d3]
+  - @leafygreen-ui/lib@6.1.2
+
+## 1.0.1
+
+### Patch Changes
+
+- 374430ea: Updates string color value to reference the same color from uiColors palette
+- 9ee1d5fc: Updates `@leafygreen-ui/lib` dependency to the most recent major version
+- Updated dependencies [c9a0d89f]
+- Updated dependencies [9ee1d5fc]
+  - @leafygreen-ui/palette@3.1.0
+  - @leafygreen-ui/lib@6.1.1
+
 ## 1.0.0
 
 ### Major Changes